refactor(#3037): relocate MonitoringStore to services layer (bucket 3)
Move the per-channel monitoring store (MonitoringStore / MonitoringEntry /
global_monitoring_store) out of server/state.rs into a new
services::monitoring_store module. It is pure in-memory state with no axum /
route dependency, so it belongs beside the rest of the service-layer infra.

server/state.rs becomes a compat facade re-exporting the same symbols, so the
crate::server::routes::state::* path (mounted via #[path] in
server/routes/mod.rs) keeps working unchanged for all route-layer and frozen
callers. Non-frozen service consumers (disk_monitor, idle_detector,
monitoring_status) are switched to the canonical crate::services::monitoring_store
path. Behavior-preserving; mirrors the established cluster/dispatch facade pattern.

Refs #3037
