+++ src/services/discord/monitoring_status.rs
@@ -13,9 +13,11 @@ use super::outbound::{
     OutboundDeduper, shared_outbound_deduper,
 };
 use super::{SharedData, health, rate_limit_wait};
-use crate::server::routes::state::{MonitoringEntry, MonitoringStore, global_monitoring_store};
 use crate::services::dispatches::discord_delivery::{
     DispatchMessagePostError, DispatchMessagePostErrorKind,
+};
+use crate::services::monitoring_store::{
+    MonitoringEntry, MonitoringStore, global_monitoring_store,
 };
 
 const RENDER_DEBOUNCE: StdDuration = StdDuration::from_millis(300);
